+++ src/colinux/arch/i386/utils.h
@@ -13,13 +13,13 @@
 #include <colinux/common/common.h>
 
 extern bool_t co_is_pae_enabled(void);
-extern unsigned long co_get_cr3(void);
-extern unsigned long co_get_cr4(void);
-extern unsigned long co_get_dr0(void);
-extern unsigned long co_get_dr1(void);
-extern unsigned long co_get_dr2(void);
-extern unsigned long co_get_dr3(void);
-extern unsigned long co_get_dr6(void);
-extern unsigned long co_get_dr7(void);
+extern uintptr_t co_get_cr3(void);
+extern uintptr_t co_get_cr4(void);
+extern uintptr_t co_get_dr0(void);
+extern uintptr_t co_get_dr1(void);
+extern uintptr_t co_get_dr2(void);
+extern uintptr_t co_get_dr3(void);
+extern uintptr_t co_get_dr6(void);
+extern uintptr_t co_get_dr7(void);
 
 #endif
