+++ src/colinux/common/libc.h
@@ -10,11 +10,11 @@
 #ifndef __COLINUX_LINUX_LIBC_H__
 #define __COLINUX_LINUX_LIBC_H__
 
-extern void *co_memset(void *s, int c, long n);
-extern void *co_memcpy(void *dest, const void *src, long n);
+extern void *co_memset(void *s, int c, int n);
+extern void *co_memcpy(void *dest, const void *src, int n);
 extern int co_memcmp(void *s1, void *, int);
-extern void *co_memmove(void *dest, const void *src, long n);
-extern void co_bzero(void *s, long n);
+extern void *co_memmove(void *dest, const void *src, int n);
+extern void co_bzero(void *s, int n);
 extern const char *co_strstr(const char *haystack, const char *needle);
 extern int co_strlen(const char *s);
 extern int co_strcmp(const char *s1, const char *s2);
